Refactor extension and create asset generator
Separates the extension.ts code in multiple files and adds an asset
generator code to provide the `tasks.json` and `launch.json` files (can
be triggered manually with a command).

- The build task generated uses the Godot CLI to build the solution (as
described in #18), I think this is the cleanest solution since we let
Godot handle the building of the project, but the user must configure
the path to the Godot executable.
- The Launch debug configuration is now created with the build
`preLaunchTask` automatically so users won't have to set it up manually
(fixes #18).
- Adds snippets to `tasks.json` to add debug configuration easily
after the file has been created.
- Adds a command to generate the `tasks.json` and `launch.json` manually
for convenience (asks the user if they want to override their current
configuration first).